This is an update of the tests using a modification of Roman Redziejowski's Java 1.8 PEG grammar; the comments had to be removed, since the pika parser can't handle them. (The copyright info is in the adjoining file.)
The pika parser was not quite able to handle that grammar, so besides the
file with comments removed (in Java.1.8.original_without_comments.peg, there
is also a modified Java.1.8.peg which loads the grammar successfuly, though
it nevertheless is unable to result in a completely sucessful parse of the
accompanying MemoTable.java file. More details in the pull request.